requirements.txt
streamlit
google-api-python-client
pandas
pymongo
SQLAlchemy
pymysql
sqlalchemy
psycopg2-binary
toml
psycopg2
psycopg2-binary